Styling fixes with user profile
