+++ Spawn-App-iOS-SwiftUI/Views/Pages/Profile/ProfileView.swift
@@ -38,9 +38,9 @@ struct ProfileView: View {
 
     // Check if this is the current user's profile
     private var isCurrentUserProfile: Bool {
-        if MockAPIService.isMocking {
+//        if MockAPIService.isMocking {
             return true
-        }
+ //       }
         guard let currentUser = userAuth.spawnUser else { return false }
         return currentUser.id == user.id
     }
@@ -55,72 +55,73 @@ struct ProfileView: View {
         NavigationStack {
             ZStack {
                 ScrollView {
-                    VStack(alignment: .center, spacing: 10) {
+                    VStack(alignment: .center, spacing: 15) {
                         // Profile Picture
-                        profilePictureSection
-                            .padding(.top, 15)
-
-                        // Name and Username - make this more reactive to changes
-                        Group {
-                            if isCurrentUserProfile,
-                                let currentUser = userAuth.spawnUser
-                            {
-                                // For the current user, always display the latest from userAuth
-                                Text(
-                                    FormatterService.shared.formatName(
-                                        user: currentUser
-                                    )
-                                )
-                                .font(.title3)
-                                .bold()
-                                .foregroundColor(universalAccentColor)
-
-                                Text("@\(currentUser.username)")
-                                    .font(.subheadline)
-                                    .foregroundColor(Color.gray)
-                                    .padding(.bottom, 5)
-                            } else {
-                                // For other users, use the passed-in user
-                                Text(
-                                    FormatterService.shared.formatName(
-                                        user: user
-                                    )
-                                )
-                                .font(.title3)
-                                .bold()
-                                .foregroundColor(universalAccentColor)
-
-                                Text("@\(user.username)")
-                                    .font(.subheadline)
-                                    .foregroundColor(Color.gray)
-                                    .padding(.bottom, 5)
-                            }
-                        }
-                        .id(refreshFlag)  // Force refresh when flag changes
+                        ProfileHeaderView(
+                            user: user,
+                            selectedImage: $selectedImage,
+                            showImagePicker: $showImagePicker,
+                            isImageLoading: $isImageLoading,
+                            isCurrentUserProfile: isCurrentUserProfile,
+                            editingState: editingState
+                        )
+                        .padding(.top, 15)
+                        
+                        // Name and Username
+                        ProfileNameView(
+                            user: user,
+                            isCurrentUserProfile: isCurrentUserProfile,
+                            userAuth: userAuth,
+                            refreshFlag: refreshFlag
+                        )
 
                         // Profile Action Buttons
-                        profileActionButtons
-                            .padding(.horizontal, 25)
-                            .padding(.bottom, 15)
+                        ProfileActionButtonsView(
+                            user: user,
+                            isCurrentUserProfile: isCurrentUserProfile,
+                            profileViewModel: profileViewModel,
+                            shareProfile: shareProfile
+                        )
+                        .padding(.horizontal, 25)
+                        .padding(.bottom, 15)
 
                         // Edit Save Cancel buttons (only when editing)
                         if isCurrentUserProfile && editingState == .save {
-                            profileEditButtons
-                                .padding(.bottom, 5)
+                            ProfileEditButtonsView(
+                                editingState: $editingState,
+                                onCancel: cancelEditing,
+                                onSave: { Task { await saveProfile() } },
+                                isImageLoading: isImageLoading
+                            )
+                            .padding(.bottom, 5)
                         }
 
                         // Interests Section with Social Media Icons
-                        interestsSection
-                            .padding(.bottom, 15)
+                        ProfileInterestsView(
+                            profileViewModel: profileViewModel, 
+                            isCurrentUserProfile: isCurrentUserProfile,
+                            editingState: editingState,
+                            whatsappLink: $whatsappLink, 
+                            instagramLink: $instagramLink,
+                            onRemoveInterest: removeInterest,
+                            openSocialMediaLink: openSocialMediaLink
+                        )
+                        .padding(.bottom, 15)
 
                         // User Stats
-                        userStatsSection
-                            .padding(.bottom, 15)
+                        ProfileStatsView(userStats: profileViewModel.userStats)
+                        .padding(.bottom, 15)
 
                         // Weekly Calendar View
-                        weeklyCalendarView
-                            .padding(.horizontal)
-                            .padding(.bottom, 15)
+                        ProfileCalendarView(
+                            profileViewModel: profileViewModel,
+                            currentMonth: $currentMonth,
+                            currentYear: $currentYear,
+                            showCalendarPopup: $showCalendarPopup,
+                            handleDaySelection: handleDaySelection
+                        )
+                        .padding(.horizontal)
+                        .padding(.bottom, 15)
                     }
                     .padding(.horizontal)
                 }
@@ -324,249 +325,21 @@ struct ProfileView: View {
             name = currentUser.name ?? ""
         }
     }
-
-    private func handleEventSelection(_ activity: CalendarActivityDTO) {
-        // First close the calendar popup
-        showCalendarPopup = false
-        
-        // Then fetch and show the event details
-        Task {
-            if let eventId = activity.eventId,
-               let _ = await profileViewModel.fetchEventDetails(eventId: eventId) {
-                await MainActor.run {
-                    showEventDetails = true
-                }
-            }
-        }
-    }
     
-    private func determineEventColor(for event: FullFeedEventDTO) -> Color {
-        // Logic to determine event color based on friend tag or category
-        if let hexCode = event.eventFriendTagColorHexCodeForRequestingUser, !hexCode.isEmpty {
-            return Color(hex: hexCode)
-        } else {
-            return event.category.color()
-        }
-    }
-}
+    private func cancelEditing() {
+        // Revert to original values from userAuth.spawnUser
+        if let currentUser = userAuth.spawnUser {
+            username = currentUser.username
+            name = currentUser.name ?? ""
+            selectedImage = nil
 
-// MARK: - Profile Picture Section
-extension ProfileView {
-    private var profilePictureSection: some View {
-        ZStack(alignment: .bottomTrailing) {
-            if isImageLoading {
-                ProgressView()
-                    .frame(width: 130, height: 130)
-            } else if let selectedImage = selectedImage {
-                Image(uiImage: selectedImage)
-                    .resizable()
-                    .scaledToFill()
-                    .frame(width: 130, height: 130)
-                    .clipShape(Circle())
-                    .transition(.opacity)
-                    .id("selectedImage-\(UUID().uuidString)")
-            } else if let profilePictureString = user.profilePicture {
-                if MockAPIService.isMocking {
-                    Image(profilePictureString)
-                        .resizable()
-                        .scaledToFill()
-                        .frame(width: 130, height: 130)
-                        .clipShape(Circle())
-                } else {
-                    AsyncImage(url: URL(string: profilePictureString)) {
-                        phase in
-                        switch phase {
-                        case .empty:
-                            ProgressView()
-                                .frame(width: 130, height: 130)
-                        case .success(let image):
-                            image
-                                .resizable()
-                                .scaledToFill()
-                                .frame(width: 130, height: 130)
-                                .clipShape(Circle())
-                                .transition(.opacity.animation(.easeInOut))
-                        case .failure:
-                            Image(systemName: "person.circle.fill")
-                                .resizable()
-                                .scaledToFit()
-                                .frame(width: 130, height: 130)
-                                .foregroundColor(Color.gray.opacity(0.5))
-                        @unknown default:
-                            Image(systemName: "person.circle.fill")
-                                .resizable()
-                                .scaledToFit()
-                                .frame(width: 130, height: 130)
-                                .foregroundColor(Color.gray.opacity(0.5))
-                        }
-                    }
-                    .id("profilePicture-\(profilePictureString)")
-                }
-            } else {
-                Image(systemName: "person.circle.fill")
-                    .resizable()
-                    .scaledToFit()
-                    .frame(width: 130, height: 130)
-                    .foregroundColor(Color.gray.opacity(0.5))
-            }
-
-            // Only show the plus button for current user's profile when in edit mode
-            if isCurrentUserProfile && editingState == .save {
-                Circle()
-                    .fill(profilePicPlusButtonColor)
-                    .frame(width: 26, height: 26)
-                    .overlay(
-                        Image(systemName: "plus")
-                            .foregroundColor(.white)
-                            .font(.system(size: 14, weight: .bold))
-                    )
-                    .offset(x: -10, y: -10)
-                    .onTapGesture {
-                        showImagePicker = true
-                    }
+            // Revert social media links
+            if let socialMedia = profileViewModel.userSocialMedia {
+                whatsappLink = socialMedia.whatsappLink ?? ""
+                instagramLink = socialMedia.instagramLink ?? ""
             }
         }
-        .animation(.easeInOut, value: selectedImage != nil)
-        .animation(.easeInOut, value: isImageLoading)
-        .sheet(
-            isPresented: $showImagePicker,
-            onDismiss: {
-                // Only show loading if we actually have a new image
-                if selectedImage != nil {
-                    DispatchQueue.main.async {
-                        isImageLoading = true
-                        DispatchQueue.main.asyncAfter(deadline: .now() + 0.3) {
-                            isImageLoading = false
-                        }
-                    }
-                }
-            }
-        ) {
-            SwiftUIImagePicker(selectedImage: $selectedImage)
-                .ignoresSafeArea()
-        }
-        .onChange(of: selectedImage) { newImage in
-            if newImage != nil {
-                // Force UI update when image changes
-                DispatchQueue.main.async {
-                    isImageLoading = true
-                    DispatchQueue.main.asyncAfter(deadline: .now() + 0.3) {
-                        isImageLoading = false
-                    }
-                }
-            }
-        }
-    }
-}
-
-// MARK: - Profile Action Buttons
-extension ProfileView {
-    private var profileActionButtons: some View {
-        HStack(spacing: 12) {
-            if isCurrentUserProfile {
-                NavigationLink(
-                    destination: EditProfileView(
-                        userId: user.id,
-                        profileViewModel: profileViewModel
-                    )
-                ) {
-                    HStack {
-                        Image(systemName: "pencil")
-                        Text("Edit Profile")
-                            .bold()
-                    }
-                    .font(.caption)
-                    .foregroundColor(universalSecondaryColor)
-                    .padding(.vertical, 24)
-                    .padding(.horizontal, 8)
-                    .frame(height: 32)
-                    .frame(maxWidth: .infinity)
-                }
-                .navigationBarBackButtonHidden(true)
-                .overlay(
-                    RoundedRectangle(cornerRadius: 12)
-                        .stroke(universalSecondaryColor, lineWidth: 1)
-                )
-            }
-
-            Button(action: {
-                shareProfile()
-            }) {
-                HStack {
-                    Image(systemName: "square.and.arrow.up")
-                    Text("Share Profile")
-                        .bold()
-                }
-                .font(.caption)
-                .foregroundColor(universalSecondaryColor)
-                .padding(.vertical, 24)
-                .padding(.horizontal, 8)
-                .frame(height: 32)
-                .frame(maxWidth: .infinity)
-                .overlay(
-                    RoundedRectangle(cornerRadius: 12)
-                        .stroke(universalSecondaryColor, lineWidth: 1)
-                )
-            }
-        }
-    }
-}
-
-// MARK: - Profile Edit Buttons
-extension ProfileView {
-    private var profileEditButtons: some View {
-        HStack(spacing: 20) {
-            // Cancel Button
-            Button(action: {
-                // Revert to original values from userAuth.spawnUser
-                if let currentUser = userAuth.spawnUser {
-                    username = currentUser.username
-                    name = currentUser.name ?? ""
-                    selectedImage = nil
-
-                    // Revert social media links
-                    if let socialMedia = profileViewModel
-                        .userSocialMedia
-                    {
-                        whatsappLink = socialMedia.whatsappLink ?? ""
-                        instagramLink = socialMedia.instagramLink ?? ""
-                    }
-                }
-                editingState = .edit
-            }) {
-                Text("Cancel")
-                    .font(.headline)
-                    .foregroundColor(universalAccentColor)
-                    .frame(maxWidth: 135)
-                    .padding()
-                    .background(
-                        RoundedRectangle(
-                            cornerRadius: universalRectangleCornerRadius
-                        )
-                        .stroke(universalAccentColor, lineWidth: 1)
-                    )
-            }
-
-            // Save Button
-            Button(action: {
-                Task {
-                    await saveProfile()
-                }
-            }) {
-                Text("Save")
-                    .font(.headline)
-                    .foregroundColor(.white)
-                    .frame(maxWidth: 135)
-                    .padding()
-                    .background(
-                        RoundedRectangle(
-                            cornerRadius: universalRectangleCornerRadius
-                        )
-                        .fill(profilePicPlusButtonColor)
-                    )
-            }
-            .disabled(isImageLoading)
-        }
+        editingState = .edit
     }
 
     private func saveProfile() async {
@@ -632,298 +405,18 @@ extension ProfileView {
             editingState = .edit
         }
     }
-}
 
-// MARK: - Interests Section
-extension ProfileView {
-    private var interestsSection: some View {
-        ZStack(alignment: .top) {
-            // Interests content
-            Group {
-                if profileViewModel.isLoadingInterests {
-                    interestsLoadingView
-                } else {
-                    interestsContentView
+    private func handleEventSelection(_ activity: CalendarActivityDTO) {
+        // First close the calendar popup
+        showCalendarPopup = false
+        
+        // Then fetch and show the event details
+        Task {
+            if let eventId = activity.eventId,
+               let _ = await profileViewModel.fetchEventDetails(eventId: eventId) {
+                await MainActor.run {
+                    showEventDetails = true
                 }
-            }
-            .padding(.top, 24)  // Add padding to push content below the header
-
-            // Position the header to be centered on the top border
-            interestsSectionHeader
-                .padding(.leading, 6)
-            //                .offset() // Align with the top border
-        }
-    }
-
-    private var interestsSectionHeader: some View {
-        HStack {
-            Text("Interests + Hobbies")
-                .font(.headline)
-                .foregroundColor(.white)
-                .padding(.vertical, 10)
-                .padding(.horizontal, 16)
-                .overlay(
-                    RoundedRectangle(cornerRadius: 8)
-                        .stroke(universalTertiaryColor, lineWidth: 1)
-                )
-                .background(universalTertiaryColor)
-                .clipShape(Capsule())
-
-            Spacer()
-
-            // Social media icons
-            if !profileViewModel.isLoadingSocialMedia {
-                socialMediaIcons
-            }
-        }
-        .padding(.horizontal)
-    }
-
-    private var socialMediaIcons: some View {
-        HStack(spacing: 10) {
-            if let whatsappLink = profileViewModel.userSocialMedia?
-                .whatsappLink, !whatsappLink.isEmpty
-            {
-                Image("whatsapp")
-                    .resizable()
-                    .scaledToFit()
-                    .frame(width: 48, height: 48)
-                    .rotationEffect(.degrees(-8))
-                    .onTapGesture {
-                        openSocialMediaLink(
-                            platform: "WhatsApp",
-                            link: whatsappLink
-                        )
-                    }
-            }
-
-            if let instagramLink = profileViewModel.userSocialMedia?
-                .instagramLink, !instagramLink.isEmpty
-            {
-                Image("instagram")
-                    .resizable()
-                    .scaledToFit()
-                    .frame(width: 40, height: 40)
-                    .rotationEffect(.degrees(8))
-                    .onTapGesture {
-                        openSocialMediaLink(
-                            platform: "Instagram",
-                            link: instagramLink
-                        )
-                    }
-            }
-        }
-    }
-
-    private var interestsLoadingView: some View {
-        ProgressView()
-            .frame(maxWidth: .infinity, alignment: .center)
-            .padding()
-    }
-
-    private var interestsContentView: some View {
-        ZStack(alignment: .topLeading) {
-            // Background for interests section
-            RoundedRectangle(cornerRadius: 15)
-                .stroke(Color.red.opacity(0.7), lineWidth: 1)
-                .background(Color.white.opacity(0.5).cornerRadius(15))
-
-            if profileViewModel.userInterests.isEmpty {
-                emptyInterestsView
-            } else {
-                // Interests as chips with proper layout
-                ScrollView {
-                    LazyVStack(alignment: .leading, spacing: 12) {
-                        // Use a simple LazyVGrid for consistent layout
-                        LazyVGrid(
-                            columns: [
-                                GridItem(
-                                    .adaptive(minimum: 80, maximum: 150),
-                                    spacing: 8
-                                )
-                            ],
-                            alignment: .leading,
-                            spacing: 8
-                        ) {
-                            ForEach(profileViewModel.userInterests, id: \.self)
-                            { interest in
-                                interestChip(interest: interest)
-                            }
-                        }
-                    }
-                    .padding()
-                }
-            }
-        }
-        .frame(height: profileViewModel.userInterests.isEmpty ? 100 : 140)
-        .padding(.horizontal)
-        .padding(.top, 5)
-    }
-
-    private var emptyInterestsView: some View {
-        Text("No interests added yet.")
-            .foregroundColor(.gray)
-            .italic()
-            .padding()
-            .padding(.top, 12)
-    }
-
-    private func interestChip(interest: String) -> some View {
-        Text(interest)
-            .font(.subheadline)
-            .padding(.vertical, 8)
-            .padding(.horizontal, 14)
-            .foregroundColor(universalAccentColor)
-            .background(Color.white)
-            .clipShape(Capsule())
-            .shadow(color: Color.black.opacity(0.1), radius: 3, x: 0, y: 2)
-            .overlay(
-                isCurrentUserProfile && editingState == .save
-                    ? HStack {
-                        Spacer()
-                        Button(action: {
-                            removeInterest(interest)
-                        }) {
-                            Image(systemName: "xmark.circle.fill")
-                                .foregroundColor(.red)
-                                .font(.caption)
-                        }
-                        .offset(x: 5, y: -8)
-                    } : nil
-            )
-    }
-}
-
-// MARK: - User Stats Section
-extension ProfileView {
-    private var userStatsSection: some View {
-        HStack(spacing: 35) {
-            VStack(spacing: 4) {
-                Image(systemName: "link")
-                    .font(.system(size: 16))
-                    .foregroundColor(.gray)
-
-                Text("\(profileViewModel.userStats?.peopleMet ?? 0)")
-                    .font(.system(size: 20, weight: .bold))
-                    .foregroundColor(universalAccentColor)
-
-                Text("People\nmet")
-                    .font(.caption2)
-                    .multilineTextAlignment(.center)
-                    .foregroundColor(.gray)
-            }
-
-            VStack(spacing: 4) {
-                Image(systemName: "star.fill")
-                    .font(.system(size: 16))
-                    .foregroundColor(.gray)
-
-                Text("\(profileViewModel.userStats?.spawnsMade ?? 0)")
-                    .font(.system(size: 20, weight: .bold))
-                    .foregroundColor(universalAccentColor)
-
-                Text("Spawns\nmade")
-                    .font(.caption2)
-                    .multilineTextAlignment(.center)
-                    .foregroundColor(.gray)
-            }
-
-            VStack(spacing: 4) {
-                Image(systemName: "calendar.badge.plus")
-                    .font(.system(size: 16))
-                    .foregroundColor(.gray)
-
-                Text("\(profileViewModel.userStats?.spawnsJoined ?? 0)")
-                    .font(.system(size: 20, weight: .bold))
-                    .foregroundColor(universalAccentColor)
-
-                Text("Spawns\njoined")
-                    .font(.caption2)
-                    .multilineTextAlignment(.center)
-                    .foregroundColor(.gray)
-            }
-        }
-    }
-}
-
-// MARK: - Weekly Calendar
-extension ProfileView {
-    private var weeklyCalendarView: some View {
-        VStack(spacing: 8) {
-            // Month and year title
-            Text(monthYearString())
-                .font(.subheadline)
-                .foregroundColor(universalAccentColor)
-                .padding(.vertical, 5)
-
-            // Days of week header
-            HStack(spacing: 0) {
-                ForEach(Array(zip(0..<weekDays.count, weekDays)), id: \.0) { index, day in
-                    Text(day)
-                        .font(.caption2)
-                        .frame(maxWidth: .infinity)
-                        .foregroundColor(.gray)
-                }
-            }
-
-            if profileViewModel.isLoadingCalendar {
-                ProgressView()
-                    .frame(maxWidth: .infinity, minHeight: 150)
-            } else {
-                // Calendar grid (clickable to show popup)
-                VStack(spacing: 6) {
-                    ForEach(0..<5, id: \.self) { row in
-                        HStack(spacing: 6) {
-                            ForEach(0..<7, id: \.self) { col in
-                                if let dayActivities = getDayActivities(row: row, col: col) {
-                                    if dayActivities.isEmpty {
-                                        // Empty day cell
-                                        RoundedRectangle(cornerRadius: 6)
-                                            .fill(Color.gray.opacity(0.2))
-                                            .frame(height: 32)
-                                    } else {
-                                        // Mini day cell with multiple activities
-                                        MiniDayCell(activities: dayActivities)
-                                            .onTapGesture {
-                                                handleDaySelection(activities: dayActivities)
-                                            }
-                                    }
-                                } else {
-                                    RoundedRectangle(cornerRadius: 6)
-                                        .fill(Color.gray.opacity(0.2))
-                                        .frame(height: 32)
-                                }
-                            }
-                        }
-                    }
-                }
-                .onTapGesture {
-                    // Load all calendar activities before showing the popup
-                    Task {
-                        await profileViewModel.fetchAllCalendarActivities()
-                        await MainActor.run {
-                            showCalendarPopup = true
-                        }
-                    }
-                }
-            }
-        }
-        .onAppear {
-            fetchCalendarData()
-        }
-        .sheet(isPresented: $showEventDetails) {
-            if let event = profileViewModel.selectedEvent {
-                // Use the same color scheme as EventCardView would
-                let eventColor = event.isSelfOwned == true ? 
-                    universalAccentColor : determineEventColor(for: event)
-                
-                EventDescriptionView(
-                    event: event,
-                    users: event.participantUsers,
-                    color: eventColor,
-                    userId: userAuth.spawnUser?.id ?? UUID()
-                )
-                .presentationDetents([.medium, .large])
             }
         }
     }
@@ -976,163 +469,19 @@ extension ProfileView {
         }
     }
     
-    // Get array of activities for a specific day cell
-    private func getDayActivities(row: Int, col: Int) -> [CalendarActivityDTO]? {
-        // Convert the original single-activity grid to an array of activities per cell
-        let activity = profileViewModel.calendarActivities[row][col]
-        
-        if activity == nil {
-            return nil
+    private func determineEventColor(for event: FullFeedEventDTO) -> Color {
+        // Logic to determine event color based on friend tag or category
+        if let hexCode = event.eventFriendTagColorHexCodeForRequestingUser, !hexCode.isEmpty {
+            return Color(hex: hexCode)
+        } else {
+            return event.category.color()
         }
-        
-        // Find all activities for this day by date checking
-        if let firstActivity = activity {
-            let day = Calendar.current.component(.day, from: firstActivity.date)
-            let month = Calendar.current.component(.month, from: firstActivity.date)
-            let year = Calendar.current.component(.year, from: firstActivity.date)
-            
-            // Filter all activities matching this date
-            return profileViewModel.allCalendarActivities.filter { act in
-                let actDay = Calendar.current.component(.day, from: act.date)
-                let actMonth = Calendar.current.component(.month, from: act.date)
-                let actYear = Calendar.current.component(.year, from: act.date)
-                
-                return actDay == day && actMonth == month && actYear == year
-            }
-        }
-        
-        return []
-    }
-
-    private var weekDays: [String] {
-        ["S", "M", "T", "W", "T", "F", "S"]
-    }
-
-    private func fetchCalendarData() {
-        Task {
-            await profileViewModel.fetchCalendarActivities(
-                month: currentMonth,
-                year: currentYear
-            )
-            // Also fetch all activities to have them ready
-            await profileViewModel.fetchAllCalendarActivities()
-        }
-    }
-
-    private func monthYearString() -> String {
-        let dateComponents = DateComponents(
-            year: currentYear,
-            month: currentMonth
-        )
-        if let date = Calendar.current.date(from: dateComponents) {
-            let formatter = DateFormatter()
-            formatter.dateFormat = "MMMM yyyy"
-            return formatter.string(from: date)
-        }
-        return "\(currentMonth)/\(currentYear)"
     }
 }
 
-// Helper struct for the mini day cell in the profile view
-struct MiniDayCell: View {
-    let activities: [CalendarActivityDTO]
-    
-    private var gradientColors: [Color] {
-        // Get up to 3 unique colors from activities
-        let colors = activities.prefix(3).compactMap { activity -> Color? in
-            if let colorHex = activity.colorHexCode, !colorHex.isEmpty {
-                return Color(hex: colorHex)
-            } else if let category = activity.eventCategory {
-                return category.color()
-            }
-            return nil
-        }
-        
-        // If no colors found, return default gray
-        if colors.isEmpty {
-            return [Color.gray.opacity(0.5)]
-        }
-        
-        // If only one color, use it with different opacity
-        if colors.count == 1 {
-            return [colors[0].opacity(0.7), colors[0].opacity(0.9)]
-        }
-        
-        return colors
-    }
-    
-    var body: some View {
-        ZStack {
-            // Gradient background
-            RoundedRectangle(cornerRadius: 6)
-                .fill(LinearGradient(
-                    gradient: Gradient(colors: gradientColors),
-                    startPoint: .topLeading,
-                    endPoint: .bottomTrailing
-                ))
-                .frame(height: 32)
-            
-            if activities.count <= 4 {
-                // Show up to 4 icons in a grid
-                let columns = [
-                    GridItem(.flexible(), spacing: 1),
-                    GridItem(.flexible(), spacing: 1)
-                ]
-                
-                LazyVGrid(columns: columns, spacing: 1) {
-                    ForEach(activities.prefix(4), id: \.id) { activity in
-                        activityIcon(for: activity)
-                            .foregroundColor(.white)
-                            .font(.system(size: 10))
-                    }
-                }
-                .padding(2)
-            } else {
-                // Show 2 icons + overflow indicator
-                HStack(spacing: 2) {
-                    ForEach(0..<2, id: \.self) { index in
-                        if index < activities.count {
-                            activityIcon(for: activities[index])
-                                .foregroundColor(.white)
-                                .font(.system(size: 10))
-                        }
-                    }
-                    
-                    Text("+\(activities.count - 2)")
-                        .font(.system(size: 10))
-                        .foregroundColor(.white)
-                        .fontWeight(.bold)
-                }
-            }
-        }
-    }
-    
-    private func activityColor(for activity: CalendarActivityDTO) -> Color {
-        // First check if activity has a custom color hex code
-        if let colorHexCode = activity.colorHexCode, !colorHexCode.isEmpty {
-            return Color(hex: colorHexCode)
-        }
-        
-        // Fallback to category color
-        guard let category = activity.eventCategory else {
-            return Color.gray.opacity(0.6) // Default color for null category
-        }
-        return category.color()
-    }
-    
-    private func activityIcon(for activity: CalendarActivityDTO) -> some View {
-        Group {
-            // If we have an icon from the backend, use it directly
-            if let icon = activity.icon, !icon.isEmpty {
-                Text(icon)
-                    .font(.system(size: 10))
-            } else {
-                // Fallback to system icon from the EventCategory enum
-                Image(systemName: activity.eventCategory?.systemIcon() ?? "star.fill")
-                    .font(.system(size: 10))
-            }
-        }
-    }
+@available(iOS 17, *)
+#Preview {
+    ProfileView(user: BaseUserDTO.danielAgapov)
 }
 
 // Extension for custom corner rounding
@@ -1154,9 +503,4 @@ struct RoundedCorner: Shape {
         )
         return Path(path.cgPath)
     }
-}
-
-@available(iOS 17, *)
-#Preview {
-    ProfileView(user: BaseUserDTO.danielAgapov)
 }
